fix(ts): create an interface for `ConversationConfiguration`
Previously clients needed to sometimes use a low-level
`ConversationConfiguration` instance, which required i.e. a `number`
for the ciphersuite instead of an enum variant, with obvious
usability problems resulting.

This creates a high-level `ConversationConfiguration` interface
that they can use instead.
